// src/openapi/gencode/models/Comment.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Json Placeholder Typicode
 * This is a partial Swagger/OpenApi modelling of the mock API available at [https://jsonplaceholder.typicode.com/](https://jsonplaceholder.typicode.com/). 
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface Comment
 */
export interface Comment {
    /**
     * 
     * @type {number}
     * @memberof Comment
     */
    id: number;
    /**
     * 
     * @type {number}
     * @memberof Comment
     */
    postId: number;
    /**
     * 
     * @type {string}
     * @memberof Comment
     */
    name?: string;
    /**
     * 
     * @type {string}
     * @memberof Comment
     */
    email?: string;
    /**
     * 
     * @type {string}
     * @memberof Comment
     */
    body: string;
}

export function CommentFromJSON(json: any): Comment {
    return CommentFromJSONTyped(json, false);
}

export function CommentFromJSONTyped(json: any, ignoreDiscriminator: boolean): Comment {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'id': json['id'],
        'postId': json['postId'],
        'name': !exists(json, 'name') ? undefined : json['name'],
        'email': !exists(json, 'email') ? undefined : json['email'],
        'body': json['body'],
    };
}

export function CommentToJSON(value?: Comment | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'id': value.id,
        'postId': value.postId,
        'name': value.name,
        'email': value.email,
        'body': value.body,
    };
}